# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChanAnalyzer API",
    "description": "A股缠论研究工作台 v2",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/scans": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Scan",
        "operationId": "create_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/limit-ups": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Limit Up Stocks",
        "operationId": "limit_up_stocks_api_v1_limit_ups_get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Limit Up Stocks Api V1 Limit Ups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/limit-ups/analyses": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Analyze Limit Ups",
        "operationId": "analyze_limit_ups_api_v1_limit_ups_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ipos": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Ipo Stocks",
        "operationId": "ipo_stocks_api_v1_ipos_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ipo Stocks Api V1 Ipos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ipos/analyses": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Analyze Ipos",
        "operationId": "analyze_ipos_api_v1_ipos_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Codes"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "codes"
        ],
        "title": "AnalysisRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobAccepted"
      },
...
      "ScanRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "buy",
              "sell",
              "hot",
              "smart"
            ],
            "title": "Mode"
          },
          "scan_side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Scan Side",
            "default": "buy"
          },
          "codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Codes"
          },
          "types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Types"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Areas"
          },
          "exclude_st": {
            "type": "boolean",
            "title": "Exclude St",
            "default": true
          },
          "rank_type": {
            "type": "string",
            "title": "Rank Type",
            "default": "top_gainers"
          },
          "top_n": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Top N",
            "default": 200
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ScanRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}